homework08/elevennote/requirements/base.txt
asgiref==3.2.10
Django==3.1.1
psycopg2==2.8.5
uWSGI==2.0.19
python-decouple==3.3
pytz==2020.1
sqlparse==0.3.1
django-widget-tweaks==1.4.8
django-wysiwyg==0.8.0
djangorestframework==3.11.1
djangorestframework-jwt==1.11.0
django-taggit==1.3.0
django-taggit-serializer==0.1.7